+++ macros/spells/thunderousSmite.js
@@ -1,0 +1,39 @@
+try {
+  if (!["mwak"].includes(args[0].item.data.actionType)) return {};
+  if (args[0].hitTargetUuids.length === 0) return {}; // did not hit anyone
+  for (let tokenUuid of args[0].hitTargetUuids) {
+    const target = await fromUuid(tokenUuid);
+    const targetActor = target.actor;
+    if (!targetActor) continue;
+    const spellDC = actor.data.flags["midi-qol"].thunderousSmite.dc;
+
+    const saveRollData =  {
+      request: "save",
+      targetUuid: target.actor.uuid,
+      ability: "str",
+      options: {
+        chatMessage: true,
+        flavor: `${CONFIG.DND5E.abilities["str"]} DC${spellDC} vs Thunderous Smite Stagger`,
+      },
+    };
+
+    // const saveRoll = await targetActor.rollAbilitySave("str", { flavor });
+    const saveRoll = await MidiQOL.socket().executeAsGM("rollAbility", saveRollData);
+
+    if (saveRoll.total < spellDC) {
+      game.dfreds.effectInterface.addEffect({ effectName: "Prone", uuid: tokenUuid });
+      ChatMessage.create({ content: `${targetActor.name} has failed the save and is pushed back 10ft and knocked prone.` });
+    }
+  }
+  Hooks.once("midi-qol.RollComplete", (workflow) => {
+    console.log("Deleting concentration");
+    const effect = MidiQOL.getConcentrationEffect(actor);
+    if (effect) effect.delete();
+    return true;
+  });
+  return { damageRoll: `2d6[thunder]`, flavor: "Thunderous Smite" };
+} catch (err) {
+  console.error(`${args[0].itemData.name} - Thunderous Smite`, err);
+}
+
+